Find subclasses more efficiently
Previous to this commit, each object in the runtime was iterated over in
order to find classes that were a subclass of a specific class.  For
example to find all subclasses of ProjectRazor::Slice, every object was
iterated over, converted to a string, the checked to see if it had the
prefix "ProjectRazor::Slice::".  This caused strange race conditions
when several requests would come in at once.

This commit introduces the Object#child_classes method. This allows it
to be called on any namespace and it will return an array of the child
classes.  For example, for the children of Project::Slice, simply
call ProjectRazor::Slice.child_classes.  This method is much more
efficient and resolves the race condition