Bug fix (taking care of issue where the get_an_ip call was returning a binary rather than a string value, prior to this change would see lines that looked like the following:
image_svc_host: !binary |-
  MTkyLjE2OC4zLjMy

at the start of the conf/razor_server.conf file (after regenerating it with a "razor config" command); without the string concatenation and the call to strip! the command returns a binary string similar to that shown above under Ruby v1.9.2 and Ruby v1.9.3 (but works without error under Ruby v1.8.7).  It has to do with the string encoding (which is UTF-8 under 1.8.7 but not under 1.9.x)