Remove repeated code and optimize case statement.
Note that internally Ruby uses === for case statement so when testing
for an object's class, "case object" is the correct way.  See

http://ruby.about.com/od/beginningruby/qt/On-Case-And-Class.htm

Previously the object under question has been converted to a string for
comparison.  This works now, but it's a little dangerous because it
assumes that the string representation of certain classes (String,
Symbol, Hash, etc.) do not change over time.